agent update

+++ src/agent.rs
@@ -56,7 +56,7 @@ fn main() {
                                         match action.as_ref() {
                                             "INPUT" => {
                                                 for arg in args.clone() {
-                                                    let res: Result<usize,_> = arg.parse();
+                                                    let res: Result<u16,_> = arg.parse();
                                                     if let Ok(num) = res {
                                                         let mut input_u: INPUT_u = unsafe { std::mem::zeroed() };
                                                         unsafe {
@@ -82,7 +82,7 @@ fn main() {
                                                 }
 
                                                 for arg in args.clone() {
-                                                    let res: Result<usize,_> = arg.parse();
+                                                    let res: Result<u16,_> = arg.parse();
                                                     if let Ok(num) = res {
                                                         let mut input_u: INPUT_u = unsafe { std::mem::zeroed() };
                                                         unsafe {
